normalize C code with gnu-indent
